requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.0
pandas==2.1.1
jupyter==1.0.0
matplotlib==3.8.0
tqdm==4.66.1
ipywidgets==8.1.1
